Common NBTCompoundGetters interface
